# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Beauty salon appointment management service",
        "title": "Beauty Salon Event Bus API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/appointments": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "appointments"
                ],
                "summary": "List appointments",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Appointment"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "appointments"
                ],
                "summary": "Create appointment",
                "parameters": [
                    {
                        "description": "appointment",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httpserver.CreateAppointmentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Appointment"
                        }
                    }
                }
            }
        },
        "/appointments/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "appointments"
                ],
                "summary": "Get appointment by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "appointment id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Appointment"
                        }
                    }
                }
            }
        },
        "/appointments/{id}/complete": {
            "post": {
                "tags": [
                    "appointments"
                ],
                "summary": "Complete appointment",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "appointment id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "definitions": {
        "httpserver.CreateAppointmentRequest": {
            "type": "object",
            "properties": {
                "client": {
                    "type": "string"
                },
                "service": {
                    "type": "string"
                }
            }
        },
        "models.Appointment": {
            "type": "object",
            "properties": {
                "client": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "service": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    }
}